feat: implement PlayerContext class
