--- src/fr/utt/lo02/tdvp/core/PlayerContext.java
@@ -1,24 +1,0 @@
-package fr.utt.lo02.tdvp.core;
-
-public class PlayerContext {
-    private Player strategy;
-
-    public PlayerContext(Player strategy) {
-    }
-
-    public void init() {
-    }
-
-    public void play() {
-    }
-
-    public String getName() {
-    }
-
-    public void setScore(int score) {
-    }
-
-    public int getScore() {
-    }
-
-}
